requirements.txt
Flask
pyserial
